src/app/exact-amount/exact-amount.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ExactAmountComponent } from './exact-amount.component';

describe('Component: ExactAmount', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ExactAmountComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([ExactAmountComponent],
      (component: ExactAmountComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ExactAmountComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(ExactAmountComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-exact-amount></app-exact-amount>
  `,
  directives: [ExactAmountComponent]
})
class ExactAmountComponentTestController {
}

